Power Functional Network Cortical Parcellation

Brain atlas for the Power cortical parcellation with 129 regions based on functional network organization of the human brain. Contains 2D polygon geometry for ggseg::geom_brain().

power()

Value

A ggseg.formats::ggseg_atlas object (cortical).

References

Power JD, Cohen AL, Nelson SM, Wig GS, Barnes KA, Church JA, Vogel AC, Laumann TO, Miezin FM, Schlaggar BL, Petersen SE (2011). Functional Network Organization of the Human Brain. Neuron, 72(4):665-678. doi:10.1016/j.neuron.2011.09.006

Examples

power()
#> 
#> ── power ggseg atlas ───────────────────────────────────────────────────────────
#> Type: cortical
#> Regions: 65
#> Hemispheres: left, right
#> Views: lateral, medial
#> Palette: ✔
#> Rendering: ✔ ggseg
#> ✔ ggseg3d (vertices)
#> ────────────────────────────────────────────────────────────────────────────────
#> # A tibble: 129 × 3
#>     hemi  region label          
#>     <chr> <chr>  <chr>          
#>   1 left  3      lh_LH_region_3 
#>   2 left  6      lh_LH_region_6 
#>   3 left  7      lh_LH_region_7 
#>   4 left  8      lh_LH_region_8 
#>   5 left  9      lh_LH_region_9 
#>   6 left  10     lh_LH_region_10
#>   7 left  11     lh_LH_region_11
#>   8 left  12     lh_LH_region_12
#>   9 left  15     lh_LH_region_15
#>  10 left  16     lh_LH_region_16
#>  11 left  17     lh_LH_region_17
#>  12 left  18     lh_LH_region_18
#>  13 left  19     lh_LH_region_19
#>  14 left  21     lh_LH_region_21
#>  15 left  22     lh_LH_region_22
#>  16 left  23     lh_LH_region_23
#>  17 left  24     lh_LH_region_24
#>  18 left  25     lh_LH_region_25
#>  19 left  26     lh_LH_region_26
#>  20 left  27     lh_LH_region_27
#>  21 left  28     lh_LH_region_28
#>  22 left  29     lh_LH_region_29
#>  23 left  30     lh_LH_region_30
#>  24 left  31     lh_LH_region_31
#>  25 left  33     lh_LH_region_33
#>  26 left  34     lh_LH_region_34
#>  27 left  35     lh_LH_region_35
#>  28 left  36     lh_LH_region_36
#>  29 left  37     lh_LH_region_37
#>  30 left  38     lh_LH_region_38
#>  31 left  39     lh_LH_region_39
#>  32 left  40     lh_LH_region_40
#>  33 left  41     lh_LH_region_41
#>  34 left  42     lh_LH_region_42
#>  35 left  43     lh_LH_region_43
#>  36 left  44     lh_LH_region_44
#>  37 left  45     lh_LH_region_45
#>  38 left  47     lh_LH_region_47
#>  39 left  48     lh_LH_region_48
#>  40 left  54     lh_LH_region_54
#>  41 left  55     lh_LH_region_55
#>  42 left  56     lh_LH_region_56
#>  43 left  58     lh_LH_region_58
#>  44 left  59     lh_LH_region_59
#>  45 left  60     lh_LH_region_60
#>  46 left  61     lh_LH_region_61
#>  47 left  62     lh_LH_region_62
#>  48 left  63     lh_LH_region_63
#>  49 left  64     lh_LH_region_64
#>  50 left  65     lh_LH_region_65
#>  51 left  1      lh_LH_region_1 
#>  52 left  2      lh_LH_region_2 
#>  53 left  4      lh_LH_region_4 
#>  54 left  5      lh_LH_region_5 
#>  55 left  13     lh_LH_region_13
#>  56 left  14     lh_LH_region_14
#>  57 left  20     lh_LH_region_20
#>  58 left  32     lh_LH_region_32
#>  59 left  46     lh_LH_region_46
#>  60 left  49     lh_LH_region_49
#>  61 left  50     lh_LH_region_50
#>  62 left  51     lh_LH_region_51
#>  63 left  52     lh_LH_region_52
#>  64 left  53     lh_LH_region_53
#>  65 left  57     lh_LH_region_57
#>  66 right 1      rh_RH_region_1 
#>  67 right 3      rh_RH_region_3 
#>  68 right 4      rh_RH_region_4 
#>  69 right 5      rh_RH_region_5 
#>  70 right 6      rh_RH_region_6 
#>  71 right 9      rh_RH_region_9 
#>  72 right 10     rh_RH_region_10
#>  73 right 13     rh_RH_region_13
#>  74 right 17     rh_RH_region_17
#>  75 right 19     rh_RH_region_19
#>  76 right 21     rh_RH_region_21
#>  77 right 22     rh_RH_region_22
#>  78 right 24     rh_RH_region_24
#>  79 right 25     rh_RH_region_25
#>  80 right 26     rh_RH_region_26
#>  81 right 33     rh_RH_region_33
#>  82 right 38     rh_RH_region_38
#>  83 right 40     rh_RH_region_40
#>  84 right 41     rh_RH_region_41
#>  85 right 44     rh_RH_region_44
#>  86 right 45     rh_RH_region_45
#>  87 right 46     rh_RH_region_46
#>  88 right 47     rh_RH_region_47
#>  89 right 52     rh_RH_region_52
#>  90 right 58     rh_RH_region_58
#>  91 right 59     rh_RH_region_59
#>  92 right 65     rh_RH_region_65
#>  93 right 7      rh_RH_region_7 
#>  94 right 8      rh_RH_region_8 
#>  95 right 11     rh_RH_region_11
#>  96 right 12     rh_RH_region_12
#>  97 right 14     rh_RH_region_14
#>  98 right 15     rh_RH_region_15
#>  99 right 16     rh_RH_region_16
#> 100 right 18     rh_RH_region_18
#> 101 right 20     rh_RH_region_20
#> 102 right 23     rh_RH_region_23
#> 103 right 27     rh_RH_region_27
#> 104 right 28     rh_RH_region_28
#> 105 right 29     rh_RH_region_29
#> 106 right 30     rh_RH_region_30
#> 107 right 31     rh_RH_region_31
#> 108 right 32     rh_RH_region_32
#> 109 right 34     rh_RH_region_34
#> 110 right 35     rh_RH_region_35
#> 111 right 36     rh_RH_region_36
#> 112 right 37     rh_RH_region_37
#> 113 right 39     rh_RH_region_39
#> 114 right 42     rh_RH_region_42
#> 115 right 43     rh_RH_region_43
#> 116 right 48     rh_RH_region_48
#> 117 right 49     rh_RH_region_49
#> 118 right 50     rh_RH_region_50
#> 119 right 51     rh_RH_region_51
#> 120 right 53     rh_RH_region_53
#> 121 right 54     rh_RH_region_54
#> 122 right 55     rh_RH_region_55
#> 123 right 56     rh_RH_region_56
#> 124 right 57     rh_RH_region_57
#> 125 right 60     rh_RH_region_60
#> 126 right 61     rh_RH_region_61
#> 127 right 62     rh_RH_region_62
#> 128 right 63     rh_RH_region_63
#> 129 right 64     rh_RH_region_64